+++ phpstan.neon.dist
@@ -33,14 +33,14 @@ includes:
 
 services:
     -
-        class: GraphQL\PhpStan\Type\Definition\Type\IsInputTypeStaticMethodTypeSpecifyingExtension
+        class: GraphQL\Tests\PhpStan\Type\Definition\Type\IsInputTypeStaticMethodTypeSpecifyingExtension
         tags:
             - phpstan.typeSpecifier.staticMethodTypeSpecifyingExtension
     -
-        class: GraphQL\PhpStan\Type\Definition\Type\IsOutputTypeStaticMethodTypeSpecifyingExtension
+        class: GraphQL\Tests\PhpStan\Type\Definition\Type\IsOutputTypeStaticMethodTypeSpecifyingExtension
         tags:
             - phpstan.typeSpecifier.staticMethodTypeSpecifyingExtension
     -
-        class: GraphQL\PhpStan\Type\Definition\Type\IsCompositeTypeStaticMethodTypeSpecifyingExtension
+        class: GraphQL\Tests\PhpStan\Type\Definition\Type\IsCompositeTypeStaticMethodTypeSpecifyingExtension
         tags:
             - phpstan.typeSpecifier.staticMethodTypeSpecifyingExtension
