+++ tests/PhpStan/Type/Definition/Type/IsCompositeTypeStaticMethodTypeSpecifyingExtension.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace GraphQL\PhpStan\Type\Definition\Type;
+namespace GraphQL\Tests\PhpStan\Type\Definition\Type;
 
-use GraphQL\Type\Definition\InputType;
-use GraphQL\Type\Definition\OutputType;
+use GraphQL\Type\Definition\CompositeType;
 use GraphQL\Type\Definition\Type;
 use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Analyser\Scope;
@@ -17,7 +16,7 @@ use PHPStan\Reflection\MethodReflection;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\StaticMethodTypeSpecifyingExtension;
 
-class IsOutputTypeStaticMethodTypeSpecifyingExtension implements StaticMethodTypeSpecifyingExtension, TypeSpecifierAwareExtension
+final class IsCompositeTypeStaticMethodTypeSpecifyingExtension implements StaticMethodTypeSpecifyingExtension, TypeSpecifierAwareExtension
 {
     /** @var TypeSpecifier */
     private $typeSpecifier;
@@ -35,11 +34,11 @@ class IsOutputTypeStaticMethodTypeSpecifyingExtension implements StaticMethodTyp
     public function isStaticMethodSupported(MethodReflection $staticMethodReflection, StaticCall $node, TypeSpecifierContext $context) : bool
     {
         // The $context argument tells us if we're in an if condition or not (as in this case).
-        return $staticMethodReflection->getName() === 'isOutputType' && ! $context->null();
+        return $staticMethodReflection->getName() === 'isCompositeType' && ! $context->null();
     }
 
     public function specifyTypes(MethodReflection $staticMethodReflection, StaticCall $node, Scope $scope, TypeSpecifierContext $context) : SpecifiedTypes
     {
-        return $this->typeSpecifier->create($node->args[0]->value, new ObjectType(OutputType::class), $context);
+        return $this->typeSpecifier->create($node->args[0]->value, new ObjectType(CompositeType::class), $context);
     }
 }
