+++ tests/PhpStan/Type/Definition/Type/IsInputTypeStaticMethodTypeSpecifyingExtension.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace GraphQL\PhpStan\Type\Definition\Type;
+namespace GraphQL\Tests\PhpStan\Type\Definition\Type;
 
-use GraphQL\Type\Definition\CompositeType;
+use GraphQL\Type\Definition\InputType;
 use GraphQL\Type\Definition\Type;
 use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Analyser\Scope;
@@ -16,7 +16,7 @@ use PHPStan\Reflection\MethodReflection;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\StaticMethodTypeSpecifyingExtension;
 
-final class IsCompositeTypeStaticMethodTypeSpecifyingExtension implements StaticMethodTypeSpecifyingExtension, TypeSpecifierAwareExtension
+class IsInputTypeStaticMethodTypeSpecifyingExtension implements StaticMethodTypeSpecifyingExtension, TypeSpecifierAwareExtension
 {
     /** @var TypeSpecifier */
     private $typeSpecifier;
@@ -34,11 +34,11 @@ final class IsCompositeTypeStaticMethodTypeSpecifyingExtension implements Static
     public function isStaticMethodSupported(MethodReflection $staticMethodReflection, StaticCall $node, TypeSpecifierContext $context) : bool
     {
         // The $context argument tells us if we're in an if condition or not (as in this case).
-        return $staticMethodReflection->getName() === 'isCompositeType' && ! $context->null();
+        return $staticMethodReflection->getName() === 'isInputType' && ! $context->null();
     }
 
     public function specifyTypes(MethodReflection $staticMethodReflection, StaticCall $node, Scope $scope, TypeSpecifierContext $context) : SpecifiedTypes
     {
-        return $this->typeSpecifier->create($node->args[0]->value, new ObjectType(CompositeType::class), $context);
+        return $this->typeSpecifier->create($node->args[0]->value, new ObjectType(InputType::class), $context);
     }
 }
